database: token page

+++ src/Database/Song.cpp
@@ -116,7 +116,7 @@ StaticPart* Song::FindPartByType(const String& type) {
 	return 0;
 }
 
-int Song::StructSuggestion::GetEstimatedDuration(int bpm) const {
+double Song::StructSuggestion::GetEstimatedDuration(int bpm) const {
 	double bars_per_min = (double)bpm / 4.0;
 	double bars_per_sec = bars_per_min / 60.0;
 	int bars = 8 * parts.GetCount();
